requirements.txt
beautifulsoup4==4.9.3
cachetools==5.2.0
certifi==2022.9.24
chardet==4.0.0
deprecation==2.1.0
googlesearch-python==1.1.0
humanize==4.4.0
idna==2.10
lxml==4.9.1
packaging==21.3
pikepdf==6.2.0
Pillow==9.2.0
pyparsing==3.0.9
PySimpleGUI==4.60.4
requests==2.25.1
soupsieve==2.3.2.post1
urllib3==1.26.12
urltitle==0.3.8
wikipedia==1.4.0
youtube-search==2.1.2
